requirements.txt
pymongo
streamlit
python-dotenv
requests
langchain
sentence_transformers
groq
huggingface-hub
langchain-community==0.3.4
langchain-core==0.3.15
langchain-experimental==0.3.2
langchain-groq==0.2.1
langchain-openai==0.2.5
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.1
numpy==1.26.4
pandas==2.2.3
pinecone==5.3.1
pinecone-client==5.0.1
scikit-learn==1.5.2
regex
tokenizers==0.20.1
torch
